Move IKE types out from internal
