+++ encr/encr_test.go
