+++ esn/esn.go
