+++ esn/esn_enable.go
