+++ integ/auth_hmac_md5_96.go
