+++ integ/auth_hmac_sha1_96.go
